accelerometer & gyro input

+++ dronex2/include/lidar.h
@@ -8,7 +8,7 @@
 
 volatile uint8_t lidar_receive[8];
 
-uint8_t lidar_transmit;
+uint16_t lidar_transmit;
 
 //void lidar_uart_init(void);
